feat : fix model estimator and add model evaluation
There was an issue regarding access to data on S3 bucket as well as parameter passing to the train.py. The parameters have been changed and the model now has access to the training data. New addition of model evalulation is added and answer to the question regarding accuracy of the model has been updated.